Adds timeout config for Confab startup
We've seen cases in which the size and diversity of the consul cluster
make it difficult for the Confab process to come up within the default
timeframe of 55 seconds. This commit allows for configuration from the
deployment manifest to be propated to the job in order to adjust that
start-up time.

Signed-off-by: Jesse Malone <jmalone@pivotal.io>